+++ lib/jade-sql/compiler.rb
@@ -28,7 +28,9 @@ require_relative 'compiler/errors'
 require_relative 'compiler/assignable'
 require_relative 'compiler/columns'
 require_relative 'compiler/selectable'
+require_relative 'compiler/selected'
 
 Jade::Extensions.register_deriver('jade-sql', JadeSql::Compiler::Assignable)
 Jade::Extensions.register_deriver('jade-sql', JadeSql::Compiler::Selectable)
 Jade::Extensions.register_check('jade-sql', :call, JadeSql::Compiler::Columns)
+Jade::Extensions.register_check('jade-sql', :call, JadeSql::Compiler::Selected)